Refactor code to remove duplicate test project and fix NotificationClient
- added unit test for GetAllNotifications
