--- src/Notify.Core.Tests/UnitTests/NotificationUnitClientTests.cs
@@ -1,548 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Net;
-using System.Net.Http;
-using System.Threading;
-using System.Threading.Tasks;
-using Moq;
-using Moq.Protected;
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
-using Notify.Client;
-using Notify.Exceptions;
-using Notify.Models;
-using Notify.Models.Responses;
-using NUnit.Framework;
-
-namespace Notify.Core.Tests.UnitTests
-{
-    [TestFixture]
-    public class NotificationUnitClientTests
-    {
-        private Mock<HttpMessageHandler> handler;
-        private NotificationClient client;
-
-        [SetUp]
-        public void SetUp()
-        {
-            handler = new Mock<HttpMessageHandler>();
-
-            var w = new HttpClientWrapper(new HttpClient(handler.Object));
-            client = new NotificationClient(w, Constants.fakeApiKey);
-        }
-
-        [TearDown]
-        public void TearDown()
-        {
-            handler = null;
-            client = null;
-        }
-
-
-        [Test, Category("Unit/NotificationClient")]
-        public void CreateNotificationClientWithInvalidApiKeyFails()
-        {
-            Assert.Throws<NotifyAuthException>(() => new NotificationClient("someinvalidkey"));
-        }
-
-
-        [Test, Category("Unit/NotificationClient")]
-        public void CreateNotificationClientWithEmptyApiKeyFails()
-        {
-            Assert.Throws<NotifyAuthException>(() => new NotificationClient(""));
-        }
-
-
-        [Test, Category("Unit/NotificationClient")]
-        public void GetNotificationWithIdCreatesExpectedRequest()
-        {
-            MockRequest(Constants.fakeNotificationJson,
-                client.GET_NOTIFICATION_URL + Constants.fakeNotificationId,
-                AssertValidRequest);
-
-            client.GetNotificationById(Constants.fakeNotificationId);
-        }
-
-
-        [Test, Category("Unit/NotificationClient")]
-        public void GetTemplateWithIdCreatesExpectedRequest()
-        {
-            MockRequest(Constants.fakeTemplateResponseJson,
-                client.GET_TEMPLATE_URL + Constants.fakeTemplateId,
-                AssertValidRequest);
-
-            client.GetTemplateByIdAndVersion(Constants.fakeTemplateId);
-        }
-
-
-        [Test, Category("Unit/NotificationClient")]
-        public void GetTemplateWithIdAndVersionCreatesExpectedRequest()
-        {
-            MockRequest(Constants.fakeTemplateResponseJson,
-                client.GET_TEMPLATE_URL + Constants.fakeTemplateId + client.VERSION_PARAM + "2",
-                AssertValidRequest);
-
-            client.GetTemplateByIdAndVersion(Constants.fakeTemplateId, 2);
-        }
-
-        [Test, Category("Unit/NotificationClient")]
-        public void GetNotificationWithIdReceivesExpectedResponse()
-        {
-            var expectedResponse = JsonConvert.DeserializeObject<Notification>(Constants.fakeNotificationJson);
-
-            MockRequest(Constants.fakeNotificationJson);
-
-            var responseNotification = client.GetNotificationById(Constants.fakeNotificationId);
-            Assert.IsTrue(expectedResponse.Equals(responseNotification));
-        }
-
-        [Test, Category("Unit/NotificationClient")]
-        public void GetTemplateWithIdReceivesExpectedResponse()
-        {
-            var expectedResponse = JsonConvert.DeserializeObject<TemplateResponse>(Constants.fakeTemplateResponseJson);
-
-            MockRequest(Constants.fakeTemplateResponseJson);
-
-            var responseTemplate = client.GetTemplateById(Constants.fakeTemplateId);
-            Assert.IsTrue(expectedResponse.Equals(responseTemplate));
-        }
-
-        [Test, Category("Unit/NotificationClient")]
-        public void GetTemplateWithIdAndVersionReceivesExpectedResponse()
-        {
-            var expectedResponse =
-                JsonConvert.DeserializeObject<TemplateResponse>(Constants.fakeTemplateResponseJson);
-
-            MockRequest(Constants.fakeTemplateResponseJson);
-
-            var responseTemplate = client.GetTemplateByIdAndVersion(Constants.fakeTemplateId, 2);
-            Assert.IsTrue(expectedResponse.Equals(responseTemplate));
-        }
-
-        [Test, Category("Unit/NotificationClient")]
-        public void GenerateTemplatePreviewGeneratesExpectedRequest()
-        {
-            Dictionary<string, dynamic> personalisation = new Dictionary<string, dynamic> {
-                    { "name", "someone" }
-            };
-
-            var o = new JObject
-            {
-                { "personalisation", JObject.FromObject(personalisation) }
-            };
-
-            MockRequest(Constants.fakeTemplatePreviewResponseJson,
-                client.GET_TEMPLATE_URL + Constants.fakeTemplateId + "/preview", AssertValidRequest, HttpMethod.Post);
-
-            var response = client.GenerateTemplatePreview(Constants.fakeTemplateId, personalisation);
-        }
-
-        [Test, Category("Unit/NotificationClient")]
-        public void GenerateTemplatePreviewReceivesExpectedResponse()
-        {
-            Dictionary<string, dynamic> personalisation = new Dictionary<string, dynamic> {
-                    { "name", "someone" }
-            };
-
-            var expected = new JObject
-            {
-                { "personalisation", JObject.FromObject(personalisation) }
-            };
-
-            MockRequest(Constants.fakeTemplatePreviewResponseJson,
-                client.GET_TEMPLATE_URL + Constants.fakeTemplateId + "/preview",
-                AssertValidRequest,
-                HttpMethod.Post,
-                AssertGetExpectedContent, expected.ToString(Formatting.None));
-
-            client.GenerateTemplatePreview(Constants.fakeTemplateId, personalisation);
-        }
-
-        [Test, Category("Unit/NotificationClient")]
-        public void GetAllTemplatesCreatesExpectedRequest()
-        {
-            MockRequest(Constants.fakeTemplateListResponseJson,
-                 client.GET_ALL_TEMPLATES_URL, AssertValidRequest);
-
-            client.GetAllTemplates();
-        }
-
-        [Test, Category("Unit/NotificationClient")]
-        public void GetAllTemplatesBySmsTypeCreatesExpectedRequest()
-        {
-            const string type = "sms";
-            MockRequest(Constants.fakeTemplateSmsListResponseJson,
-                         client.GET_ALL_TEMPLATES_URL+ client.TYPE_PARAM + type, AssertValidRequest);
-
-            client.GetAllTemplates(type);
-        }
-
-        [Test, Category("Unit/NotificationClient")]
-        public void GetAllTemplatesByEmailTypeCreatesExpectedRequest()
-        {
-            const string type = "email";
-
-            MockRequest(Constants.fakeTemplateEmailListResponseJson,
-                         client.GET_ALL_TEMPLATES_URL+ client.TYPE_PARAM + type, AssertValidRequest);
-
-            client.GetAllTemplates(type);
-        }
-
-        [Test, Category("Unit/NotificationClient")]
-        public void GetAllTemplatesForEmptyListReceivesExpectedResponse()
-        {
-            var expectedResponse = JsonConvert.DeserializeObject<TemplateList>(Constants.fakeTemplateEmptyListResponseJson);
-
-               MockRequest(Constants.fakeTemplateEmptyListResponseJson);
-
-            TemplateList templateList = client.GetAllTemplates();
-
-            List<TemplateResponse> templates = templateList.templates;
-
-            Assert.IsTrue(templates.Count == 0);
-        }
-
-        [Test, Category("Unit/NotificationClient")]
-        public void GetAllTemplatesReceivesExpectedResponse()
-        {
-            TemplateList expectedResponse = JsonConvert.DeserializeObject<TemplateList>(Constants.fakeTemplateListResponseJson);
-
-            MockRequest(Constants.fakeTemplateListResponseJson);
-
-            TemplateList templateList = client.GetAllTemplates();
-
-            List<TemplateResponse> templates = templateList.templates;
-
-            Assert.AreEqual(templates.Count, expectedResponse.templates.Count);
-            for (int i = 0; i < templates.Count; i++)
-            {
-                Assert.IsTrue(expectedResponse.templates[i].Equals(templates[i]));
-            }
-        }
-
-        [Test, Category("Unit/NotificationClient")]
-        public void GetAllTemplatesBySmsTypeReceivesExpectedResponse()
-        {
-            const string type = "sms";
-
-            TemplateList expectedResponse =
-                JsonConvert.DeserializeObject<TemplateList>(Constants.fakeTemplateSmsListResponseJson);
-
-            MockRequest(Constants.fakeTemplateSmsListResponseJson,
-                         client.GET_ALL_TEMPLATES_URL + client.TYPE_PARAM + type, AssertValidRequest);
-
-            TemplateList templateList = client.GetAllTemplates(type);
-
-            List<TemplateResponse> templates = templateList.templates;
-
-            Assert.AreEqual(templates.Count, expectedResponse.templates.Count);
-            for (int i = 0; i < templates.Count; i++)
-            {
-                Assert.IsTrue(expectedResponse.templates[i].Equals(templates[i]));
-            }
-        }
-
-        [Test, Category("Unit/NotificationClient")]
-        public void GetAllTemplatesByEmailTypeReceivesExpectedResponse()
-        {
-            const string type = "email";
-
-            TemplateList expectedResponse =
-                JsonConvert.DeserializeObject<TemplateList>(Constants.fakeTemplateEmailListResponseJson);
-
-            MockRequest(Constants.fakeTemplateEmailListResponseJson,
-                         client.GET_ALL_TEMPLATES_URL + client.TYPE_PARAM + type, AssertValidRequest);
-
-            TemplateList templateList = client.GetAllTemplates(type);
-
-            List<TemplateResponse> templates = templateList.templates;
-
-            Assert.AreEqual(templates.Count, expectedResponse.templates.Count);
-            for (int i = 0; i < templates.Count; i++)
-            {
-                Assert.IsTrue(expectedResponse.templates[i].Equals(templates[i]));
-            }
-        }
-
-        [Test, Category("Unit/NotificationClient")]
-        public void GetAllReceivedTextsCreatesExpectedRequest()
-        {
-            MockRequest(Constants.fakeReceivedTextListResponseJson,
-                 client.GET_RECEIVED_TEXTS_URL, AssertValidRequest);
-
-            client.GetReceivedTexts();
-        }
-
-        [Test, Category("Unit/NotificationClient")]
-        public void GetAllReceivedTextsReceivesExpectedResponse()
-        {
-            MockRequest(Constants.fakeReceivedTextListResponseJson,
-                 client.GET_RECEIVED_TEXTS_URL, AssertValidRequest);
-
-            ReceivedTextListResponse expectedResponse =
-                JsonConvert.DeserializeObject<ReceivedTextListResponse>(Constants.fakeReceivedTextListResponseJson);
-
-            MockRequest(Constants.fakeReceivedTextListResponseJson,
-                         client.GET_RECEIVED_TEXTS_URL, AssertValidRequest);
-
-            ReceivedTextListResponse receivedTextList = client.GetReceivedTexts();
-
-            List<ReceivedTextResponse> receivedTexts = receivedTextList.receivedTexts;
-
-            Assert.AreEqual(receivedTexts.Count, expectedResponse.receivedTexts.Count);
-            for (int i = 0; i < receivedTexts.Count; i++)
-            {
-                Assert.IsTrue(expectedResponse.receivedTexts[i].Equals(receivedTexts[i]));
-            }
-        }
-
-        [Test, Category("Unit/NotificationClient")]
-        public void SendSmsNotificationGeneratesExpectedRequest()
-        {
-            Dictionary<string, dynamic> personalisation = new Dictionary<string, dynamic>
-                {
-                    { "name", "someone" }
-                };
-            JObject expected = new JObject
-            {
-                { "phone_number", Constants.fakePhoneNumber },
-                { "template_id", Constants.fakeTemplateId },
-                { "personalisation", JObject.FromObject(personalisation) }
-            };
-
-            MockRequest(Constants.fakeSmsNotificationResponseJson,
-                client.SEND_SMS_NOTIFICATION_URL,
-                AssertValidRequest,
-                HttpMethod.Post,
-                AssertGetExpectedContent, expected.ToString(Formatting.None));
-
-            SmsNotificationResponse response = client.SendSms(Constants.fakePhoneNumber, Constants.fakeTemplateId, personalisation);
-        }
-
-        [Test, Category("Unit/NotificationClient")]
-        public void SendSmsNotificationGeneratesExpectedResponse()
-        {
-            Dictionary<string, dynamic> personalisation = new Dictionary<string, dynamic>
-                {
-                    { "name", "someone" }
-                };
-            SmsNotificationResponse expectedResponse = JsonConvert.DeserializeObject<SmsNotificationResponse>(Constants.fakeSmsNotificationResponseJson);
-
-            MockRequest(Constants.fakeSmsNotificationResponseJson);
-
-            SmsNotificationResponse actualResponse = client.SendSms(Constants.fakePhoneNumber, Constants.fakeTemplateId, personalisation);
-
-            Assert.IsTrue(expectedResponse.Equals(actualResponse));
-        }
-
-        [Test, Category("Unit/NotificationClient")]
-        public void SendEmailNotificationGeneratesExpectedRequest()
-        {
-            Dictionary<string, dynamic> personalisation = new Dictionary<string, dynamic>
-                {
-                    { "name", "someone" }
-                };
-            JObject expected = new JObject
-            {
-                { "email_address", Constants.fakeEmail },
-                { "template_id", Constants.fakeTemplateId },
-                { "personalisation", JObject.FromObject(personalisation) },
-                { "reference", Constants.fakeNotificationReference }
-            };
-
-            MockRequest(Constants.fakeTemplatePreviewResponseJson,
-                client.SEND_EMAIL_NOTIFICATION_URL,
-                AssertValidRequest,
-                HttpMethod.Post,
-                AssertGetExpectedContent, expected.ToString(Formatting.None));
-
-            EmailNotificationResponse response = client.SendEmail(Constants.fakeEmail, Constants.fakeTemplateId, personalisation, Constants.fakeNotificationReference);
-        }
-
-        [Test, Category("Unit/NotificationClient")]
-        public void SendEmailNotificationGeneratesExpectedResponse()
-        {
-            Dictionary<string, dynamic> personalisation = new Dictionary<string, dynamic>
-                {
-                    { "name", "someone" }
-                };
-            EmailNotificationResponse expectedResponse = JsonConvert.DeserializeObject<EmailNotificationResponse>(Constants.fakeEmailNotificationResponseJson);
-
-            MockRequest(Constants.fakeEmailNotificationResponseJson);
-
-            EmailNotificationResponse actualResponse = client.SendEmail(Constants.fakeEmail, Constants.fakeTemplateId, personalisation, Constants.fakeNotificationReference);
-
-            Assert.IsTrue(expectedResponse.Equals(actualResponse));
-
-        }
-
-        [Test, Category("Unit/NotificationClient")]
-        public void SendLetterNotificationGeneratesExpectedRequest()
-        {
-            Dictionary<string, dynamic> personalisation = new Dictionary<string, dynamic>
-                {
-                    { "address_line_1", "Foo" },
-                    { "address_line_2", "Bar" },
-                    { "postcode", "Baz" }
-                };
-            JObject expected = new JObject
-            {
-                { "template_id", Constants.fakeTemplateId },
-                { "personalisation", JObject.FromObject(personalisation) },
-                { "reference", Constants.fakeNotificationReference }
-            };
-
-            MockRequest(Constants.fakeTemplatePreviewResponseJson,
-                client.SEND_LETTER_NOTIFICATION_URL,
-                AssertValidRequest,
-                HttpMethod.Post,
-                AssertGetExpectedContent, expected.ToString(Formatting.None));
-
-            LetterNotificationResponse response = client.SendLetter(Constants.fakeTemplateId, personalisation, Constants.fakeNotificationReference);
-        }
-
-        [Test, Category("Unit/NotificationClient")]
-        public void SendLetterNotificationGeneratesExpectedResponse()
-        {
-            Dictionary<string, dynamic> personalisation = new Dictionary<string, dynamic>
-                {
-                    { "address_line_1", "Foo" },
-                    { "address_line_2", "Bar" },
-                    { "postcode", "Baz" }
-                };
-            LetterNotificationResponse expectedResponse = JsonConvert.DeserializeObject<LetterNotificationResponse>(Constants.fakeLetterNotificationResponseJson);
-
-            MockRequest(Constants.fakeLetterNotificationResponseJson);
-
-            LetterNotificationResponse actualResponse = client.SendLetter(Constants.fakeTemplateId, personalisation, Constants.fakeNotificationReference);
-
-            Assert.IsTrue(expectedResponse.Equals(actualResponse));
-
-        }
-
-        private static void AssertGetExpectedContent(string expected, string content)
-        {
-            Assert.IsNotNull(content);
-            Assert.AreEqual(expected, content);
-        }
-
-        private void AssertValidRequest(string uri, HttpRequestMessage r, HttpMethod method = null)
-        {
-            if (method == null)
-            {
-                method = HttpMethod.Get;
-            }
-                
-            Assert.AreEqual(r.Method, method);
-            Assert.AreEqual(r.RequestUri.ToString(), client.BaseUrl + uri);
-            Assert.IsNotNull(r.Headers.Authorization);
-            Assert.IsNotNull(r.Headers.UserAgent);
-            Assert.AreEqual(r.Headers.UserAgent.ToString(), client.GetUserAgent());
-            Assert.AreEqual(r.Headers.Accept.ToString(), "application/json");
-        }
-
-        private void MockRequest(string content, string uri,
-                          Action<string, HttpRequestMessage, HttpMethod> _assertValidRequest = null,
-                          HttpMethod method = null,
-                          Action<string, string> _assertGetExpectedContent = null, string expected = null)
-        {
-            handler.Protected()
-                .Setup<Task<HttpResponseMessage>>("SendAsync", ItExpr.IsAny<HttpRequestMessage>(), ItExpr.IsAny<CancellationToken>())
-                .Returns(Task<HttpResponseMessage>.Factory.StartNew(() => new HttpResponseMessage
-                {
-                    StatusCode = HttpStatusCode.OK,
-                    Content = new StringContent(content)
-                }))
-                .Callback<HttpRequestMessage, CancellationToken>((r, c) =>
-                {
-                    _assertValidRequest(uri, r, method);
-
-                    if (r.Content == null || _assertGetExpectedContent == null) return;
-
-                    var response = r.Content.ReadAsStringAsync().Result;
-                    _assertGetExpectedContent(expected, response);
-                });
-        }
-
-        private void MockRequest(string content)
-        {
-
-            handler.Protected()
-                .Setup<Task<HttpResponseMessage>>("SendAsync", ItExpr.IsAny<HttpRequestMessage>(), ItExpr.IsAny<CancellationToken>())
-                .Returns(Task<HttpResponseMessage>.Factory.StartNew(() => new HttpResponseMessage
-                {
-                    StatusCode = HttpStatusCode.OK,
-                    Content = new StringContent(content)
-                }));
-        }
-
-        [Test, Category("Unit/NotificationClient")]
-        public void SendEmailNotificationWithReplyToIdGeneratesExpectedRequest()
-        {
-            var personalisation = new Dictionary<string, dynamic>
-            {
-                { "name", "someone" }
-            };
-
-            var expected = new JObject
-            {
-                { "email_address", Constants.fakeEmail },
-                { "template_id", Constants.fakeTemplateId },
-                { "personalisation", JObject.FromObject(personalisation) },
-                { "reference", Constants.fakeNotificationReference },
-                { "email_reply_to_id", Constants.fakeReplyToId}
-            };
-
-            MockRequest(Constants.fakeTemplateEmailListResponseJson,
-                client.SEND_EMAIL_NOTIFICATION_URL,
-                AssertValidRequest,
-                HttpMethod.Post,
-                AssertGetExpectedContent,
-                expected.ToString(Formatting.None));
-
-            var response = client.SendEmail(Constants.fakeEmail, Constants.fakeTemplateId, personalisation, Constants.fakeNotificationReference, Constants.fakeReplyToId);
-        }
-
-        [Test, Category("Unit/NotificationClient")]
-        public void SendEmailNotificationWithReplyToIdGeneratesExpectedResponse()
-        {
-            var personalisation = new Dictionary<string, dynamic>
-            {
-                { "name", "someone" }
-            };
-
-            var expectedResponse = JsonConvert.DeserializeObject<EmailNotificationResponse>(Constants.fakeEmailNotificationResponseJson);
-
-            MockRequest(Constants.fakeEmailNotificationResponseJson);
-            
-            var actualResponse = client.SendEmail(Constants.fakeEmail, Constants.fakeTemplateId, personalisation, Constants.fakeNotificationReference, Constants.fakeReplyToId);
-
-            Assert.IsTrue(expectedResponse.Equals(actualResponse));
-        }
-
-        [Test, Category("Unit/NotificationClient")]
-        public void SendSmsNotificationWithSmsSenderIdGeneratesExpectedRequest()
-        {
-            var personalisation = new Dictionary<string, dynamic>
-                {
-                    { "name", "someone" }
-                };
-            var expected = new JObject
-            {
-                { "phone_number", Constants.fakePhoneNumber },
-                { "template_id", Constants.fakeTemplateId },
-                { "personalisation", JObject.FromObject(personalisation) },
-                { "sms_sender_id", Constants.fakeSMSSenderId }
-            };
-
-            MockRequest(Constants.fakeSmsNotificationWithSMSSenderIdResponseJson,
-                client.SEND_SMS_NOTIFICATION_URL,
-                AssertValidRequest,
-                HttpMethod.Post,
-                AssertGetExpectedContent, expected.ToString(Formatting.None));
-
-            var response = client.SendSms(
-                Constants.fakePhoneNumber, Constants.fakeTemplateId, personalisation: personalisation, smsSenderId: Constants.fakeSMSSenderId);
-        }
-    }
-}
