+++ src/Notify.Tests/UnitTests/Constants.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Notify.UnitTests
+namespace Notify.Tests.UnitTests
 {
     public static class Constants
     {
@@ -43,6 +43,223 @@ namespace Notify.UnitTests
             }
         }
 
+        public static String fakeNotificationsJson
+        {
+            get
+            {
+                return @"{
+                    ""notifications"": [
+                        {
+                            ""completed_at"": null,
+                            ""created_at"": ""2016-11-22T11:21:13.133522Z"",
+                            ""email_address"": null,
+                            ""id"": ""902e4312-bc4a-4c87-8c3e-9f4144ca36fd"",
+                            ""line_1"": null,
+                            ""line_2"": null,
+                            ""line_3"": null,
+                            ""line_4"": null,
+                            ""line_5"": null,
+                            ""line_6"": null,
+                            ""phone_number"": ""+447588767647"",
+                            ""postcode"": null,
+                            ""reference"": null,
+                            ""sent_at"": ""2016-11-22T16:16:09.885808Z"",
+                            ""status"": ""sending"",
+                            ""template"": {
+                                ""id"": ""913e9fa6-9cbb-44ad-8f58-38487dccfd82"",
+                                ""uri"": ""/service/fd29e421-24b6-4f45-ac5c-e642a1d68641/template/323e9fa6-9cbb-44ad-8f64-38487dccfd43"",
+                                ""version"": 2  },
+                            ""type"": ""sms""
+                        },
+                        {
+                            ""completed_at"": null,
+                            ""created_at"": ""2016-11-23T11:21:13.133522Z"",
+                            ""email_address"": ""someone@example.com"",
+                            ""id"": ""902e4312-bc4a-4c87-8c3e-9f4144ca36fd"",
+                            ""line_1"": null,
+                            ""line_2"": null,
+                            ""line_3"": null,
+                            ""line_4"": null,
+                            ""line_5"": null,
+                            ""line_6"": null,
+                            ""phone_number"": null,
+                            ""postcode"": null,
+                            ""reference"": null,
+                            ""sent_at"": ""2016-11-23T16:16:09.885808Z"",
+                            ""status"": ""sending"",
+                            ""template"": {
+                                ""id"": ""913e9fa6-9cbb-44ad-8f58-38487dccfd84"",
+                                ""uri"": ""/service/fd29e421-24b6-4f45-ac5c-e642a1d68641/template/323e9fa6-9cbb-44ad-8f64-38487dccfd43"",
+                                ""version"": 2  },
+                            ""type"": ""email""
+                        }
+                    ]
+                }";
+            }
+        }
+        public static String fakeSmsNotificationsJson
+        {
+            get
+            {
+                return @"{
+                    ""notifications"": [
+                        {
+                            ""completed_at"": null,
+                            ""created_at"": ""2016-11-22T11:21:13.133522Z"",
+                            ""email_address"": null,
+                            ""id"": ""902e4312-bc4a-4c87-8c3e-9f4144ca36fd"",
+                            ""line_1"": null,
+                            ""line_2"": null,
+                            ""line_3"": null,
+                            ""line_4"": null,
+                            ""line_5"": null,
+                            ""line_6"": null,
+                            ""phone_number"": ""+447588767647"",
+                            ""postcode"": null,
+                            ""reference"": null,
+                            ""sent_at"": ""2016-11-22T16:16:09.885808Z"",
+                            ""status"": ""sending"",
+                            ""template"": {
+                                ""id"": ""913e9fa6-9cbb-44ad-8f58-38487dccfd82"",
+                                ""uri"": ""/service/fd29e421-24b6-4f45-ac5c-e642a1d68641/template/323e9fa6-9cbb-44ad-8f64-38487dccfd43"",
+                                ""version"": 2  },
+                            ""type"": ""sms""
+                        },
+                        {
+                            ""completed_at"": null,
+                            ""created_at"": ""2016-11-24T11:21:13.133522Z"",
+                            ""email_address"": null,
+                            ""id"": ""902e4312-bc4a-4c87-8c3e-9f4144ca36fd"",
+                            ""line_1"": null,
+                            ""line_2"": null,
+                            ""line_3"": null,
+                            ""line_4"": null,
+                            ""line_5"": null,
+                            ""line_6"": null,
+                            ""phone_number"": ""+447588767647"",
+                            ""postcode"": null,
+                            ""reference"": null,
+                            ""sent_at"": ""2016-11-24T16:16:09.885808Z"",
+                            ""status"": ""sending"",
+                            ""template"": {
+                                ""id"": ""913e9fa6-9cbb-44ad-8f58-38487dccfd84"",
+                                ""uri"": ""/service/fd29e421-24b6-4f45-ac5c-e642a1d68641/template/323e9fa6-9cbb-44ad-8f64-38487dccfd43"",
+                                ""version"": 2  },
+                            ""type"": ""sms""
+                        }
+                    ]
+                }";
+            }
+        }
+
+        public static String fakeEmailNotificationsJson
+        {
+            get
+            {
+                return @"{
+                    ""notifications"": [
+                        {
+                            ""completed_at"": null,
+                            ""created_at"": ""2016-11-22T11:21:13.133522Z"",
+                            ""email_address"": ""someone@example.com"",
+                            ""id"": ""902e4312-bc4a-4c87-8c3e-9f4144ca36fd"",
+                            ""line_1"": null,
+                            ""line_2"": null,
+                            ""line_3"": null,
+                            ""line_4"": null,
+                            ""line_5"": null,
+                            ""line_6"": null,
+                            ""phone_number"": null,
+                            ""postcode"": null,
+                            ""reference"": null,
+                            ""sent_at"": ""2016-11-22T16:16:09.885808Z"",
+                            ""status"": ""sending"",
+                            ""template"": {
+                                ""id"": ""913e9fa6-9cbb-44ad-8f58-38487dccfd82"",
+                                ""uri"": ""/service/fd29e421-24b6-4f45-ac5c-e642a1d68641/template/323e9fa6-9cbb-44ad-8f64-38487dccfd43"",
+                                ""version"": 2  },
+                            ""type"": ""email""
+                        },
+                        {
+                            ""completed_at"": null,
+                            ""created_at"": ""2016-11-24T11:21:13.133522Z"",
+                            ""email_address"": ""someone2@example.com"",
+                            ""id"": ""902e4312-bc4a-4c87-8c3e-9f4144ca36fd"",
+                            ""line_1"": null,
+                            ""line_2"": null,
+                            ""line_3"": null,
+                            ""line_4"": null,
+                            ""line_5"": null,
+                            ""line_6"": null,
+                            ""phone_number"": null,
+                            ""postcode"": null,
+                            ""reference"": null,
+                            ""sent_at"": ""2016-11-24T16:16:09.885808Z"",
+                            ""status"": ""sending"",
+                            ""template"": {
+                                ""id"": ""913e9fa6-9cbb-44ad-8f58-38487dccfd84"",
+                                ""uri"": ""/service/fd29e421-24b6-4f45-ac5c-e642a1d68641/template/323e9fa6-9cbb-44ad-8f64-38487dccfd43"",
+                                ""version"": 2  },
+                            ""type"": ""email""
+                        }
+                    ]
+                }";
+            }
+        }
+        public static String fakeLetterNotificationsJson
+        {
+            get
+            {
+                return @"{
+                    ""notifications"": [
+                        {
+                            ""completed_at"": null,
+                            ""created_at"": ""2016-11-22T11:21:13.133522Z"",
+                            ""email_address"": null,
+                            ""id"": ""902e4312-bc4a-4c87-8c3e-9f4144ca36fd"",
+                            ""line_1"": ""Mr Someone"",
+                            ""line_2"": ""1 Test Street"",
+                            ""line_3"": null,
+                            ""line_4"": null,
+                            ""line_5"": null,
+                            ""line_6"": null,
+                            ""phone_number"": null,
+                            ""postcode"": ""N1"",
+                            ""reference"": null,
+                            ""sent_at"": ""2016-11-22T16:16:09.885808Z"",
+                            ""status"": ""sending"",
+                            ""template"": {
+                                ""id"": ""913e9fa6-9cbb-44ad-8f58-38487dccfd82"",
+                                ""uri"": ""/service/fd29e421-24b6-4f45-ac5c-e642a1d68641/template/323e9fa6-9cbb-44ad-8f64-38487dccfd43"",
+                                ""version"": 2  },
+                            ""type"": ""letter""
+                        },
+                        {
+                            ""completed_at"": null,
+                            ""created_at"": ""2016-11-24T11:21:13.133522Z"",
+                            ""email_address"": null,
+                            ""id"": ""902e4312-bc4a-4c87-8c3e-9f4144ca36fd"",
+                            ""line_1"": ""Mrs Somebody"",
+                            ""line_2"": ""2 Test Street"",
+                            ""line_3"": null,
+                            ""line_4"": null,
+                            ""line_5"": null,
+                            ""line_6"": null,
+                            ""phone_number"": null,
+                            ""postcode"": ""N1"",
+                            ""reference"": null,
+                            ""sent_at"": ""2016-11-24T16:16:09.885808Z"",
+                            ""status"": ""sending"",
+                            ""template"": {
+                                ""id"": ""913e9fa6-9cbb-44ad-8f58-38487dccfd84"",
+                                ""uri"": ""/service/fd29e421-24b6-4f45-ac5c-e642a1d68641/template/323e9fa6-9cbb-44ad-8f64-38487dccfd43"",
+                                ""version"": 2  },
+                            ""type"": ""letter""
+                        }
+                    ]
+                }";
+            }
+        }
         public static String fakeTemplateResponseJson { get {
                 return @"{
                             ""updated_at"": null,
