+++ src/Notify.Tests/UnitTests/NotificationUnitClientTests.cs
@@ -12,7 +12,6 @@ using Notify.Client;
 using Notify.Exceptions;
 using Notify.Models;
 using Notify.Models.Responses;
-using Notify.UnitTests;
 using NUnit.Framework;
 
 namespace Notify.Tests.UnitTests
@@ -20,8 +19,8 @@ namespace Notify.Tests.UnitTests
     [TestFixture]
     public class NotificationUnitClientTests
     {
-        Mock<HttpMessageHandler> handler;
-        NotificationClient client;
+        private Mock<HttpMessageHandler> handler;
+        private NotificationClient client;
 
         [SetUp]
         public void SetUp()
@@ -39,13 +38,11 @@ namespace Notify.Tests.UnitTests
             client = null;
         }
 
-
         [Test, Category("Unit/NotificationClient")]
         public void CreateNotificationClientWithInvalidApiKeyFails()
         {
             Assert.Throws<NotifyAuthException>(() => new NotificationClient("someinvalidkey"));
         }
-
 
         [Test, Category("Unit/NotificationClient")]
         public void CreateNotificationClientWithEmptyApiKeyFails()
@@ -53,33 +50,100 @@ namespace Notify.Tests.UnitTests
             Assert.Throws<NotifyAuthException>(() => new NotificationClient(""));
         }
 
-
         [Test, Category("Unit/NotificationClient")]
         public void GetNotificationWithIdCreatesExpectedRequest()
         {
-            mockRequest(Constants.fakeNotificationJson,
+            MockRequest(Constants.fakeNotificationJson,
                 client.GET_NOTIFICATION_URL + Constants.fakeNotificationId,
                 AssertValidRequest);
 
             client.GetNotificationById(Constants.fakeNotificationId);
         }
 
+        [Test, Category("Unit/NotificationClient")]
+        public void GetAllNotificationsCreatesExpectedResult()
+        {
+            MockRequest(Constants.fakeNotificationsJson,
+                client.GET_ALL_NOTIFICATIONS_URL,
+                AssertValidRequest);
+
+            client.GetNotifications();
+        }
+
+        [Test, Category("Unit/NotificationClient")]
+        public void GetAllNotificationsWithStatusCreatesExpectedResult()
+        {
+            MockRequest(Constants.fakeNotificationsJson,
+                client.GET_ALL_NOTIFICATIONS_URL + "?status=sending",
+                AssertValidRequest);
+
+            client.GetNotifications(status: "sending");
+        }
+
+        [Test, Category("Unit/NotificationClient")]
+        public void GetAllNotificationsWithReferenceCreatesExpectedResult()
+        {
+            MockRequest(Constants.fakeNotificationsJson,
+                client.GET_ALL_NOTIFICATIONS_URL + "?reference=foo",
+                AssertValidRequest);
+
+            client.GetNotifications(reference: "foo");
+        }
+
+        [Test, Category("Unit/NotificationClient")]
+        public void GetAllSmsNotificationsWithStatusAndReferenceWithCreatesExpectedResult()
+        {
+            MockRequest(Constants.fakeNotificationsJson,
+                client.GET_ALL_NOTIFICATIONS_URL + "?template_type=sms&status=sending&reference=foo",
+                AssertValidRequest);
+
+            client.GetNotifications(templateType: "sms", status: "sending", reference: "foo");
+        }
+
+        [Test, Category("Unit/NotificationClient")]
+        public void GetAllSmsNotificationsCreatesExpectedResult()
+        {
+            MockRequest(Constants.fakeSmsNotificationResponseJson,
+                client.GET_ALL_NOTIFICATIONS_URL + "?template_type=sms",
+                AssertValidRequest);
+
+            client.GetNotifications(templateType: "sms");
+        }
+
+        [Test, Category("Unit/NotificationClient")]
+        public void GetAllEmailNotificationsCreatesExpectedResult()
+        {
+            MockRequest(Constants.fakeEmailNotificationResponseJson,
+                client.GET_ALL_NOTIFICATIONS_URL + "?template_type=email",
+                AssertValidRequest);
+
+            client.GetNotifications(templateType: "email");
+        }
+
+        [Test, Category("Unit/NotificationClient")]
+        public void GetAllLetterNotificationsCreatesExpectedResult()
+        {
+            MockRequest(Constants.fakeEmailNotificationResponseJson,
+                client.GET_ALL_NOTIFICATIONS_URL + "?template_type=letter",
+                AssertValidRequest);
+
+            client.GetNotifications(templateType: "letter");
+        }
 
         [Test, Category("Unit/NotificationClient")]
         public void GetTemplateWithIdCreatesExpectedRequest()
         {
-            mockRequest(Constants.fakeTemplateResponseJson,
+            MockRequest(Constants.fakeTemplateResponseJson,
                 client.GET_TEMPLATE_URL + Constants.fakeTemplateId,
                 AssertValidRequest);
 
             client.GetTemplateByIdAndVersion(Constants.fakeTemplateId);
         }
 
-
         [Test, Category("Unit/NotificationClient")]
         public void GetTemplateWithIdAndVersionCreatesExpectedRequest()
         {
-            mockRequest(Constants.fakeTemplateResponseJson,
+            MockRequest(Constants.fakeTemplateResponseJson,
                 client.GET_TEMPLATE_URL + Constants.fakeTemplateId + client.VERSION_PARAM + "2",
                 AssertValidRequest);
 
@@ -91,7 +155,7 @@ namespace Notify.Tests.UnitTests
         {
             var expectedResponse = JsonConvert.DeserializeObject<Notification>(Constants.fakeNotificationJson);
 
-            mockRequest(Constants.fakeNotificationJson);
+            MockRequest(Constants.fakeNotificationJson);
 
             var responseNotification = client.GetNotificationById(Constants.fakeNotificationId);
             Assert.IsTrue(expectedResponse.Equals(responseNotification));
@@ -102,7 +166,7 @@ namespace Notify.Tests.UnitTests
         {
             var expectedResponse = JsonConvert.DeserializeObject<TemplateResponse>(Constants.fakeTemplateResponseJson);
 
-            mockRequest(Constants.fakeTemplateResponseJson);
+            MockRequest(Constants.fakeTemplateResponseJson);
 
             var responseTemplate = client.GetTemplateById(Constants.fakeTemplateId);
             Assert.IsTrue(expectedResponse.Equals(responseTemplate));
@@ -114,7 +178,7 @@ namespace Notify.Tests.UnitTests
             var expectedResponse =
                 JsonConvert.DeserializeObject<TemplateResponse>(Constants.fakeTemplateResponseJson);
 
-            mockRequest(Constants.fakeTemplateResponseJson);
+            MockRequest(Constants.fakeTemplateResponseJson);
 
             var responseTemplate = client.GetTemplateByIdAndVersion(Constants.fakeTemplateId, 2);
             Assert.IsTrue(expectedResponse.Equals(responseTemplate));
@@ -123,7 +187,7 @@ namespace Notify.Tests.UnitTests
         [Test, Category("Unit/NotificationClient")]
         public void GenerateTemplatePreviewGeneratesExpectedRequest()
         {
-            Dictionary<String, dynamic> personalisation = new Dictionary<String, dynamic> {
+            Dictionary<string, dynamic> personalisation = new Dictionary<string, dynamic> {
                     { "name", "someone" }
             };
 
@@ -132,7 +196,7 @@ namespace Notify.Tests.UnitTests
                 { "personalisation", JObject.FromObject(personalisation) }
             };
 
-            mockRequest(Constants.fakeTemplatePreviewResponseJson,
+            MockRequest(Constants.fakeTemplatePreviewResponseJson,
                 client.GET_TEMPLATE_URL + Constants.fakeTemplateId + "/preview", AssertValidRequest, HttpMethod.Post);
 
             var response = client.GenerateTemplatePreview(Constants.fakeTemplateId, personalisation);
@@ -141,7 +205,7 @@ namespace Notify.Tests.UnitTests
         [Test, Category("Unit/NotificationClient")]
         public void GenerateTemplatePreviewReceivesExpectedResponse()
         {
-            Dictionary<String, dynamic> personalisation = new Dictionary<String, dynamic> {
+            Dictionary<string, dynamic> personalisation = new Dictionary<string, dynamic> {
                     { "name", "someone" }
             };
 
@@ -150,7 +214,7 @@ namespace Notify.Tests.UnitTests
                 { "personalisation", JObject.FromObject(personalisation) }
             };
 
-            mockRequest(Constants.fakeTemplatePreviewResponseJson,
+            MockRequest(Constants.fakeTemplatePreviewResponseJson,
                 client.GET_TEMPLATE_URL + Constants.fakeTemplateId + "/preview",
                 AssertValidRequest,
                 HttpMethod.Post,
@@ -162,7 +226,7 @@ namespace Notify.Tests.UnitTests
         [Test, Category("Unit/NotificationClient")]
         public void GetAllTemplatesCreatesExpectedRequest()
         {
-            mockRequest(Constants.fakeTemplateListResponseJson,
+            MockRequest(Constants.fakeTemplateListResponseJson,
                  client.GET_ALL_TEMPLATES_URL, AssertValidRequest);
 
             client.GetAllTemplates();
@@ -171,8 +235,8 @@ namespace Notify.Tests.UnitTests
         [Test, Category("Unit/NotificationClient")]
         public void GetAllTemplatesBySmsTypeCreatesExpectedRequest()
         {
-            const String type = "sms";
-            mockRequest(Constants.fakeTemplateSmsListResponseJson,
+            const string type = "sms";
+            MockRequest(Constants.fakeTemplateSmsListResponseJson,
                          client.GET_ALL_TEMPLATES_URL+ client.TYPE_PARAM + type, AssertValidRequest);
 
             client.GetAllTemplates(type);
@@ -181,9 +245,9 @@ namespace Notify.Tests.UnitTests
         [Test, Category("Unit/NotificationClient")]
         public void GetAllTemplatesByEmailTypeCreatesExpectedRequest()
         {
-            const String type = "email";
+            const string type = "email";
 
-            mockRequest(Constants.fakeTemplateEmailListResponseJson,
+            MockRequest(Constants.fakeTemplateEmailListResponseJson,
                          client.GET_ALL_TEMPLATES_URL+ client.TYPE_PARAM + type, AssertValidRequest);
 
             client.GetAllTemplates(type);
@@ -194,7 +258,7 @@ namespace Notify.Tests.UnitTests
         {
             var expectedResponse = JsonConvert.DeserializeObject<TemplateList>(Constants.fakeTemplateEmptyListResponseJson);
 
-               mockRequest(Constants.fakeTemplateEmptyListResponseJson);
+               MockRequest(Constants.fakeTemplateEmptyListResponseJson);
 
             TemplateList templateList = client.GetAllTemplates();
 
@@ -208,7 +272,7 @@ namespace Notify.Tests.UnitTests
         {
             TemplateList expectedResponse = JsonConvert.DeserializeObject<TemplateList>(Constants.fakeTemplateListResponseJson);
 
-            mockRequest(Constants.fakeTemplateListResponseJson);
+            MockRequest(Constants.fakeTemplateListResponseJson);
 
             TemplateList templateList = client.GetAllTemplates();
 
@@ -224,12 +288,12 @@ namespace Notify.Tests.UnitTests
         [Test, Category("Unit/NotificationClient")]
         public void GetAllTemplatesBySmsTypeReceivesExpectedResponse()
         {
-            const String type = "sms";
+            const string type = "sms";
 
             TemplateList expectedResponse =
                 JsonConvert.DeserializeObject<TemplateList>(Constants.fakeTemplateSmsListResponseJson);
 
-            mockRequest(Constants.fakeTemplateSmsListResponseJson,
+            MockRequest(Constants.fakeTemplateSmsListResponseJson,
                          client.GET_ALL_TEMPLATES_URL + client.TYPE_PARAM + type, AssertValidRequest);
 
             TemplateList templateList = client.GetAllTemplates(type);
@@ -246,12 +310,12 @@ namespace Notify.Tests.UnitTests
         [Test, Category("Unit/NotificationClient")]
         public void GetAllTemplatesByEmailTypeReceivesExpectedResponse()
         {
-            const String type = "email";
+            const string type = "email";
 
             TemplateList expectedResponse =
                 JsonConvert.DeserializeObject<TemplateList>(Constants.fakeTemplateEmailListResponseJson);
 
-            mockRequest(Constants.fakeTemplateEmailListResponseJson,
+            MockRequest(Constants.fakeTemplateEmailListResponseJson,
                          client.GET_ALL_TEMPLATES_URL + client.TYPE_PARAM + type, AssertValidRequest);
 
             TemplateList templateList = client.GetAllTemplates(type);
@@ -268,7 +332,7 @@ namespace Notify.Tests.UnitTests
         [Test, Category("Unit/NotificationClient")]
         public void GetAllReceivedTextsCreatesExpectedRequest()
         {
-            mockRequest(Constants.fakeReceivedTextListResponseJson,
+            MockRequest(Constants.fakeReceivedTextListResponseJson,
                  client.GET_RECEIVED_TEXTS_URL, AssertValidRequest);
 
             client.GetReceivedTexts();
@@ -277,13 +341,13 @@ namespace Notify.Tests.UnitTests
         [Test, Category("Unit/NotificationClient")]
         public void GetAllReceivedTextsReceivesExpectedResponse()
         {
-            mockRequest(Constants.fakeReceivedTextListResponseJson,
+            MockRequest(Constants.fakeReceivedTextListResponseJson,
                  client.GET_RECEIVED_TEXTS_URL, AssertValidRequest);
 
             ReceivedTextListResponse expectedResponse =
                 JsonConvert.DeserializeObject<ReceivedTextListResponse>(Constants.fakeReceivedTextListResponseJson);
 
-            mockRequest(Constants.fakeReceivedTextListResponseJson,
+            MockRequest(Constants.fakeReceivedTextListResponseJson,
                          client.GET_RECEIVED_TEXTS_URL, AssertValidRequest);
 
             ReceivedTextListResponse receivedTextList = client.GetReceivedTexts();
@@ -300,7 +364,7 @@ namespace Notify.Tests.UnitTests
         [Test, Category("Unit/NotificationClient")]
         public void SendSmsNotificationGeneratesExpectedRequest()
         {
-            Dictionary<String, dynamic> personalisation = new Dictionary<String, dynamic>
+            Dictionary<string, dynamic> personalisation = new Dictionary<string, dynamic>
                 {
                     { "name", "someone" }
                 };
@@ -311,7 +375,7 @@ namespace Notify.Tests.UnitTests
                 { "personalisation", JObject.FromObject(personalisation) }
             };
 
-            mockRequest(Constants.fakeSmsNotificationResponseJson,
+            MockRequest(Constants.fakeSmsNotificationResponseJson,
                 client.SEND_SMS_NOTIFICATION_URL,
                 AssertValidRequest,
                 HttpMethod.Post,
@@ -323,13 +387,13 @@ namespace Notify.Tests.UnitTests
         [Test, Category("Unit/NotificationClient")]
         public void SendSmsNotificationGeneratesExpectedResponse()
         {
-            Dictionary<String, dynamic> personalisation = new Dictionary<String, dynamic>
+            Dictionary<string, dynamic> personalisation = new Dictionary<string, dynamic>
                 {
                     { "name", "someone" }
                 };
             SmsNotificationResponse expectedResponse = JsonConvert.DeserializeObject<SmsNotificationResponse>(Constants.fakeSmsNotificationResponseJson);
 
-            mockRequest(Constants.fakeSmsNotificationResponseJson);
+            MockRequest(Constants.fakeSmsNotificationResponseJson);
 
             SmsNotificationResponse actualResponse = client.SendSms(Constants.fakePhoneNumber, Constants.fakeTemplateId, personalisation);
 
@@ -339,7 +403,7 @@ namespace Notify.Tests.UnitTests
         [Test, Category("Unit/NotificationClient")]
         public void SendEmailNotificationGeneratesExpectedRequest()
         {
-            Dictionary<String, dynamic> personalisation = new Dictionary<String, dynamic>
+            Dictionary<string, dynamic> personalisation = new Dictionary<string, dynamic>
                 {
                     { "name", "someone" }
                 };
@@ -351,7 +415,7 @@ namespace Notify.Tests.UnitTests
                 { "reference", Constants.fakeNotificationReference }
             };
 
-            mockRequest(Constants.fakeTemplatePreviewResponseJson,
+            MockRequest(Constants.fakeTemplatePreviewResponseJson,
                 client.SEND_EMAIL_NOTIFICATION_URL,
                 AssertValidRequest,
                 HttpMethod.Post,
@@ -363,13 +427,13 @@ namespace Notify.Tests.UnitTests
         [Test, Category("Unit/NotificationClient")]
         public void SendEmailNotificationGeneratesExpectedResponse()
         {
-            Dictionary<String, dynamic> personalisation = new Dictionary<String, dynamic>
+            Dictionary<string, dynamic> personalisation = new Dictionary<string, dynamic>
                 {
                     { "name", "someone" }
                 };
             EmailNotificationResponse expectedResponse = JsonConvert.DeserializeObject<EmailNotificationResponse>(Constants.fakeEmailNotificationResponseJson);
 
-            mockRequest(Constants.fakeEmailNotificationResponseJson);
+            MockRequest(Constants.fakeEmailNotificationResponseJson);
 
             EmailNotificationResponse actualResponse = client.SendEmail(Constants.fakeEmail, Constants.fakeTemplateId, personalisation, Constants.fakeNotificationReference);
 
@@ -380,13 +444,12 @@ namespace Notify.Tests.UnitTests
         [Test, Category("Unit/NotificationClient")]
         public void SendLetterNotificationGeneratesExpectedRequest()
         {
-            Dictionary<String, dynamic> personalisation = new Dictionary<String, dynamic>
-            {
-                { "address_line_1", "Foo" },
-                { "address_line_2", "Bar" },
-                { "postcode", "Baz" }
-            };
-
+            Dictionary<string, dynamic> personalisation = new Dictionary<string, dynamic>
+                {
+                    { "address_line_1", "Foo" },
+                    { "address_line_2", "Bar" },
+                    { "postcode", "Baz" }
+                };
             JObject expected = new JObject
             {
                 { "template_id", Constants.fakeTemplateId },
@@ -394,7 +457,7 @@ namespace Notify.Tests.UnitTests
                 { "reference", Constants.fakeNotificationReference }
             };
 
-            mockRequest(Constants.fakeTemplatePreviewResponseJson,
+            MockRequest(Constants.fakeTemplatePreviewResponseJson,
                 client.SEND_LETTER_NOTIFICATION_URL,
                 AssertValidRequest,
                 HttpMethod.Post,
@@ -406,16 +469,15 @@ namespace Notify.Tests.UnitTests
         [Test, Category("Unit/NotificationClient")]
         public void SendLetterNotificationGeneratesExpectedResponse()
         {
-            Dictionary<String, dynamic> personalisation = new Dictionary<String, dynamic>
-            {
-                { "address_line_1", "Foo" },
-                { "address_line_2", "Bar" },
-                { "postcode", "Baz" }
-            };
-
+            Dictionary<string, dynamic> personalisation = new Dictionary<string, dynamic>
+                {
+                    { "address_line_1", "Foo" },
+                    { "address_line_2", "Bar" },
+                    { "postcode", "Baz" }
+                };
             LetterNotificationResponse expectedResponse = JsonConvert.DeserializeObject<LetterNotificationResponse>(Constants.fakeLetterNotificationResponseJson);
 
-            mockRequest(Constants.fakeLetterNotificationResponseJson);
+            MockRequest(Constants.fakeLetterNotificationResponseJson);
 
             LetterNotificationResponse actualResponse = client.SendLetter(Constants.fakeTemplateId, personalisation, Constants.fakeNotificationReference);
 
@@ -423,16 +485,19 @@ namespace Notify.Tests.UnitTests
 
         }
 
-        void AssertGetExpectedContent(String expected, String content)
+        private static void AssertGetExpectedContent(string expected, string content)
         {
             Assert.IsNotNull(content);
             Assert.AreEqual(expected, content);
         }
 
-        void AssertValidRequest(String uri, HttpRequestMessage r, HttpMethod method = null)
+        private void AssertValidRequest(string uri, HttpRequestMessage r, HttpMethod method = null)
         {
             if (method == null)
+            {
                 method = HttpMethod.Get;
+            }
+                
             Assert.AreEqual(r.Method, method);
             Assert.AreEqual(r.RequestUri.ToString(), client.BaseUrl + uri);
             Assert.IsNotNull(r.Headers.Authorization);
@@ -441,57 +506,50 @@ namespace Notify.Tests.UnitTests
             Assert.AreEqual(r.Headers.Accept.ToString(), "application/json");
         }
 
-        void mockRequest(String content, String uri,
-                          Action<String, HttpRequestMessage, HttpMethod> _assertValidRequest = null,
+        private void MockRequest(string content, string uri,
+                          Action<string, HttpRequestMessage, HttpMethod> _assertValidRequest = null,
                           HttpMethod method = null,
-                          Action<String, String> _assertGetExpectedContent = null, String expected = null)
+                          Action<string, string> _assertGetExpectedContent = null, string expected = null)
         {
             handler.Protected()
                 .Setup<Task<HttpResponseMessage>>("SendAsync", ItExpr.IsAny<HttpRequestMessage>(), ItExpr.IsAny<CancellationToken>())
-                .Returns(Task<HttpResponseMessage>.Factory.StartNew(() =>
+                .Returns(Task<HttpResponseMessage>.Factory.StartNew(() => new HttpResponseMessage
                 {
-                    return new HttpResponseMessage
-                    {
-                        StatusCode = HttpStatusCode.OK,
-                        Content = new StringContent(content)
-                    };
+                    StatusCode = HttpStatusCode.OK,
+                    Content = new StringContent(content)
                 }))
                 .Callback<HttpRequestMessage, CancellationToken>((r, c) =>
                 {
                     _assertValidRequest(uri, r, method);
 
-                    if (r.Content != null && _assertGetExpectedContent != null)
-                    {
-                        String response = r.Content.ReadAsStringAsync().Result;
-                        _assertGetExpectedContent(expected, response);
-                    }
+                    if (r.Content == null || _assertGetExpectedContent == null) return;
+
+                    var response = r.Content.ReadAsStringAsync().Result;
+                    _assertGetExpectedContent(expected, response);
                 });
         }
 
-        void mockRequest(String content)
+        private void MockRequest(string content)
         {
 
             handler.Protected()
                 .Setup<Task<HttpResponseMessage>>("SendAsync", ItExpr.IsAny<HttpRequestMessage>(), ItExpr.IsAny<CancellationToken>())
-                .Returns(Task<HttpResponseMessage>.Factory.StartNew(() =>
+                .Returns(Task<HttpResponseMessage>.Factory.StartNew(() => new HttpResponseMessage
                 {
-                    return new HttpResponseMessage
-                    {
-                        StatusCode = HttpStatusCode.OK,
-                        Content = new StringContent(content)
-                    };
+                    StatusCode = HttpStatusCode.OK,
+                    Content = new StringContent(content)
                 }));
         }
 
         [Test, Category("Unit/NotificationClient")]
         public void SendEmailNotificationWithReplyToIdGeneratesExpectedRequest()
         {
-            Dictionary<String, dynamic> personalisation = new Dictionary<String, dynamic>
+            var personalisation = new Dictionary<string, dynamic>
             {
                 { "name", "someone" }
             };
 
-            JObject expected = new JObject
+            var expected = new JObject
             {
                 { "email_address", Constants.fakeEmail },
                 { "template_id", Constants.fakeTemplateId },
@@ -500,29 +558,29 @@ namespace Notify.Tests.UnitTests
                 { "email_reply_to_id", Constants.fakeReplyToId}
             };
 
-            mockRequest(Constants.fakeTemplateEmailListResponseJson,
+            MockRequest(Constants.fakeTemplateEmailListResponseJson,
                 client.SEND_EMAIL_NOTIFICATION_URL,
                 AssertValidRequest,
                 HttpMethod.Post,
                 AssertGetExpectedContent,
                 expected.ToString(Formatting.None));
 
-            EmailNotificationResponse response = client.SendEmail(Constants.fakeEmail, Constants.fakeTemplateId, personalisation, Constants.fakeNotificationReference, Constants.fakeReplyToId);
+            var response = client.SendEmail(Constants.fakeEmail, Constants.fakeTemplateId, personalisation, Constants.fakeNotificationReference, Constants.fakeReplyToId);
         }
 
         [Test, Category("Unit/NotificationClient")]
         public void SendEmailNotificationWithReplyToIdGeneratesExpectedResponse()
         {
-            Dictionary<String, dynamic> personalisation = new Dictionary<String, dynamic>
+            var personalisation = new Dictionary<string, dynamic>
             {
                 { "name", "someone" }
             };
 
-            EmailNotificationResponse expectedResponse = JsonConvert.DeserializeObject<EmailNotificationResponse>(Constants.fakeEmailNotificationResponseJson);
+            var expectedResponse = JsonConvert.DeserializeObject<EmailNotificationResponse>(Constants.fakeEmailNotificationResponseJson);
 
-            mockRequest(Constants.fakeEmailNotificationResponseJson);
+            MockRequest(Constants.fakeEmailNotificationResponseJson);
             
-            EmailNotificationResponse actualResponse = client.SendEmail(Constants.fakeEmail, Constants.fakeTemplateId, personalisation, Constants.fakeNotificationReference, Constants.fakeReplyToId);
+            var actualResponse = client.SendEmail(Constants.fakeEmail, Constants.fakeTemplateId, personalisation, Constants.fakeNotificationReference, Constants.fakeReplyToId);
 
             Assert.IsTrue(expectedResponse.Equals(actualResponse));
         }
@@ -530,11 +588,11 @@ namespace Notify.Tests.UnitTests
         [Test, Category("Unit/NotificationClient")]
         public void SendSmsNotificationWithSmsSenderIdGeneratesExpectedRequest()
         {
-            Dictionary<String, dynamic> personalisation = new Dictionary<String, dynamic>
+            var personalisation = new Dictionary<string, dynamic>
                 {
                     { "name", "someone" }
                 };
-            JObject expected = new JObject
+            var expected = new JObject
             {
                 { "phone_number", Constants.fakePhoneNumber },
                 { "template_id", Constants.fakeTemplateId },
@@ -542,13 +600,13 @@ namespace Notify.Tests.UnitTests
                 { "sms_sender_id", Constants.fakeSMSSenderId }
             };
 
-            mockRequest(Constants.fakeSmsNotificationWithSMSSenderIdResponseJson,
+            MockRequest(Constants.fakeSmsNotificationWithSMSSenderIdResponseJson,
                 client.SEND_SMS_NOTIFICATION_URL,
                 AssertValidRequest,
                 HttpMethod.Post,
                 AssertGetExpectedContent, expected.ToString(Formatting.None));
 
-            SmsNotificationResponse response = client.SendSms(
+            var response = client.SendSms(
                 Constants.fakePhoneNumber, Constants.fakeTemplateId, personalisation: personalisation, smsSenderId: Constants.fakeSMSSenderId);
         }
     }
