--- src/Notify.Tests/packages.config
@@ -1,20 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Castle.Core" version="4.2.1" targetFramework="net461" />
-  <package id="JWT" version="3.1.1" targetFramework="net461" />
-  <package id="Microsoft.NETCore.Platforms" version="1.1.0" targetFramework="net461" />
-  <package id="Microsoft.Win32.Primitives" version="4.0.1" targetFramework="net461" />
-  <package id="Moq" version="4.8.1" targetFramework="net461" />
-  <package id="MSTest.TestAdapter" version="1.1.18" targetFramework="net461" />
-  <package id="MSTest.TestFramework" version="1.1.18" targetFramework="net461" />
-  <package id="NETStandard.Library" version="2.0.1" targetFramework="net461" />
-  <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net461" />
-  <package id="NUnit" version="3.9.0" targetFramework="net461" />
-  <package id="System.Net.Http" version="4.3.3" targetFramework="net461" />
-  <package id="System.Security.Cryptography.Algorithms" version="4.3.0" targetFramework="net461" />
-  <package id="System.Security.Cryptography.Encoding" version="4.3.0" targetFramework="net461" />
-  <package id="System.Security.Cryptography.Primitives" version="4.3.0" targetFramework="net461" />
-  <package id="System.Security.Cryptography.X509Certificates" version="4.3.0" targetFramework="net461" />
-  <package id="System.Threading.Tasks.Extensions" version="4.3.0" targetFramework="net461" />
-  <package id="System.ValueTuple" version="4.3.0" targetFramework="net461" />
-</packages>
