+++ src/Notify/Notify.csproj
@@ -1,7 +1,17 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
+    <RestoreProjectStyle>PackageReference</RestoreProjectStyle>
     <TargetFrameworks>netstandard2.0;net46;net461</TargetFrameworks>
+    <Version>1.7.0</Version>
+    <Authors>GDS</Authors>
+    <Description>Net Notifications Client</Description>
+    <PackageRequireLicenseAcceptance>false</PackageRequireLicenseAcceptance>
+    <PackageReleaseNotes>Netstandard release for core 2.0, framework 4.6, 4.6.1</PackageReleaseNotes>
+  </PropertyGroup>
+
+  <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Release|netstandard2.0|AnyCPU'">
+    <DocumentationFile>bin\Release\netstandard2.0\Notify.xml</DocumentationFile>
   </PropertyGroup>
 
   <ItemGroup>
